checking if the blog about to be deleted exist

+++ src/controllers/blogController.js
@@ -111,6 +111,13 @@ class blogController {
   static async deleteBlog(req, res) {
     const id = req.params.id
 
+    if (!await blogService.findOneBlog(id)) {
+      return res.status(404).json({
+        status: 404,
+        error: 'no blog with that id found'
+      })
+    }
+
     try {
       await blogService.deleteB(id)
       res.json({
